Adding hw_power_validation.csv

+++ accelwattch_hw_profiler/collate_power.sh
@@ -3,6 +3,12 @@
 SCRIPT_DIR=`pwd`
 
 output_folder="collated_power"
+
+if [ ! "${1}" == "validation_power_reports" ] && [ ! "${1}" == "ubench_power_reports" ]; then
+    echo "Please enter a correct power reports directory, one of: [ubench_power_reports,validation_power_reports]. For example: ./collate_power.sh validation_power_reports"
+	exit 1
+fi
+
 if [ -e "$SCRIPT_DIR/${1}" ] && [ -d "$SCRIPT_DIR/${1}" ]; then
 	if [ -d "$output_folder" ]; then
 		rm -r "$output_folder"
@@ -17,6 +23,11 @@ if [ -e "$SCRIPT_DIR/${1}" ] && [ -d "$SCRIPT_DIR/${1}" ]; then
 		done
 	done
 	python gen_hw_power_csv.py $output_folder
+    if [  "${1}" == "validation_power_reports" ]; then
+        mv hw_power_results.csv hw_power_validation.csv
+    else
+        mv hw_power_results.csv hw_power_ubench.csv
+    fi
 else
 	echo "Please enter a correct power reports directory. Example: ./collate_power.sh validation_power_reports"
 	exit 1
